+++ git-keeper-robot/gkeeprobot/keywords/ServerSetupKeywords.py
@@ -11,7 +11,7 @@ class ServerSetupKeywords:
             line = 'Professor,Doctor,{}@gitkeeper.edu'.format(faculty_name)
             control.run_vm_python_script('keeper', 'add_to_file.py', 'faculty.csv', line)
 
-    def add_file(self, username, filename, target_filename):
+    def add_file_to_server(self, username, filename, target_filename):
         control.copy(username, filename, target_filename)
 
     def start_gkeepd(self):
