+++ tests/acceptance/files/malformed_cs1.csv
@@ -1,0 +1,3 @@
+Last,First,name@gitkeeper.edu
+Bad,Line
+Last,First,bar@gitkeeper.edu
